requirements.txt
Scrapy==2.11.2
psycopg2-binary==2.9.9
twilio==9.3.2
python-dotenv==1.0.1
streamlit==1.38.0
schedule==1.2.2